style: 不要ファイルを削除
